==> newFlowCohortExtractionPipeline/MGB/icd_pos/1_icd_filter.ipynb <==
# Goal of this notebook:
 - Filter notes by ICD code indicating SAH:
   - ICD +
   - ICD -
 - Plan to do so:
   - For ICD +, filter using regex to include only ICD codes that I have listed
   - For ICD -, do the opposite
   - Remember that there are more rows than unique patients.  ie: some patients will have multiple ICD codes all linked to the same note
   - To fix that - try to condense rows into one per note by making the ICD column contain a list of ICD's (not tidy but I feel like it works here)

### ICD +
 - Filter through the icd data by the 'ICDCD' column with codes that match ^(I60|430).*
 - I will be looking at all ICD codes assigned +/- ! month the date of the discharge summary
 - Check the format of the dates
 - Merge based on patient ID and date

# Overall Plan
1 - All ICD+/- assignments separated. 
- Relavent info icd csv contains: BDSPPatientID, DateAssigned ('ShiftedContactDTS'), ICD ('ICDCD')
- Relavent info notes df contains: BDSPPatientID, NoteType ('NoteTypeFull'), DateWritten ('CreateDate'), NoteTextFileName ('DeidentifiedName')

# imports
import pandas as pd
from thunderpack import ThunderReader
from tqdm import tqdm
import re

reader = ThunderReader('/home/jsearle/bigDrive/Dropbox/zz_EHR_Thunderpacks/MGB/thunderpack_icd_9_10_1m_MGB')
key_length = len(list(reader.keys()))
print(key_length)

# filter by below icd codes
code_regex = '^(I60|430)'
dfs = []
for i in tqdm(range(1, key_length + 1)):
    df = reader[f'ICD_partition_{i}']
    df = df[df['ICDCD'].astype(str).str.match(code_regex, flags = re.I)]
    dfs.append(df)

# show num of total SAH icd codes given
# show example of df
filtered_icd_df = pd.concat(dfs, axis=0, ignore_index=True)
print(len(filtered_icd_df))
filtered_icd_df.head()

# clean up df, keep only relevant info
keepColumns = ['BDSPPatientID', 'ShiftedContactDTS', 'ICDCD']
clean_icd_df = filtered_icd_df[keepColumns]
clean_icd_df.head()

# rename columns
rename_dict = { 
    'ShiftedContactDTS': 'DateICD', 
    'ICDCD': 'ICD', 
}

clean_icd_df = clean_icd_df.rename(columns=rename_dict)

clean_icd_df.head()


# save df as csv
clean_icd_df.to_csv('1_icd_plus_df.csv', index=False)

### ICD -
 - Filter through the icd data by the 'ICDCD' column with codes that *don't* match ^(I60|430).*
 - I will be looking at all ICD codes assigned +/- ! month the date of the discharge summary
 - Check the format of the dates
 - Merge based on patient ID and date
