==> newFlowCohortExtractionPipeline/MGB/icd_pos/2_discharge_summary_filter.ipynb <==
# imports
import pandas as pd
from thunderpack import ThunderReader
from tqdm import tqdm
import re

### Discharge Summaries
 - Read in all note metadata CSV files and filter InpatientNoteTypeDSC column for only 'Discharge Summary'
 - Keep track of: Note title, patient ID, date

# get note metadata for each year
# filter by discharge summary type
startYear = 2013
dfs = []
for i in range(0, 12):
    df = pd.read_csv(f'/home/jsearle/bigDrive/Dropbox/zz_EHR_Thunderpacks/MGB/MGB_Deidentified_Notes_March12th2024/mgb_notes_{startYear + i}_metadata.csv')
    df = df[df['InpatientNoteTypeDSC'] == 'Discharge Summary']
    dfs.append(df)

# show dataframe, number of discharge summaries

discharge_summaries_df = pd.concat(dfs, axis=0, ignore_index=True)
numDischargeSummaries = len(discharge_summaries_df)
print(numDischargeSummaries)
discharge_summaries_df.head()

# clean up df, keep only relevant info
keepColumns = ['BDSPPatientID', 'ContactDate',  'DeidentifiedName']
discharge_summaries_df = discharge_summaries_df[keepColumns]
discharge_summaries_df.head()

# rename columns
rename_dict = { 
    'ContactDate': 'DateNote', 
    'DeidentifiedName': 'NoteTextFile', 
}

discharge_summaries_df = discharge_summaries_df.rename(columns=rename_dict)

# change date column from str to datetime
discharge_summaries_df.loc[:, 'DateNote'] = pd.to_datetime(discharge_summaries_df['DateNote'], format='%Y%m%d')
discharge_summaries_df.head()

# read in icd filtered df
icd_filter_df = pd.read_csv('1_icd_plus_df.csv')
# change column from str to datetime
icd_filter_df['DateICD'] = pd.to_datetime(icd_filter_df['DateICD'])

# merge based on patient id
ICD_plus_DS_df = pd.merge(icd_filter_df, discharge_summaries_df, on='BDSPPatientID', how='inner')
ICD_plus_DS_df.head()


# filter for DateICD +/- 1 month of DateNote

from datetime import timedelta


time_window = timedelta(days=30)

# Filter rows where DateICD is within 1 month before or after DateNote
filtered_df = ICD_plus_DS_df[abs(ICD_plus_DS_df['DateICD'] - ICD_plus_DS_df['DateNote']) <= time_window]

print(len(filtered_df))
filtered_df.head()




# save as csv
filtered_df.to_csv('2_discharge_summaries_df.csv', index=False)
