==> newFlowCohortExtractionPipeline/MGB/icd_pos/3_cpt_filter.ipynb <==
# imports
import pandas as pd
from thunderpack import ThunderReader
from tqdm import tqdm
import re

# Goal of this notebook:
 - Filter my current dataframe (1_icd_plus_df.csv) to only include patients who have either had a head CT or a brain MRI
 - Plan to do so:
   - Read in cpt df
   - Filter for head CT and brain MRI codes
   - Extract list of patient ID's
   - Read in 1_icd_plus_df.csv
   - Pull ID column, compare with cpt ID's, and keep overlap
   - Filter 1_icd_plus_df based on the new list

reader = ThunderReader('/home/jsearle/bigDrive/Dropbox/zz_EHR_Thunderpacks/MGB/thunderpack_cpt_1m_MGB')
key_length = len(list(reader.keys()))
print(key_length)

### CPT Codes to filter for:
**70450**: 	CT Head or Brain without contrast\
**70460**:	CT Head or Brain with contrast\
**70470**: 	CT Head or Brain without contrast, followed by contrast and further sections\
**70551**: 	MRI Brain (including brain stem) without contrast\
**70552**: 	MRI Brain (including brain stem) with contrast\
**70553**: 	MRI Brain (including brain stem) without contrast, followed by with contrast

# get note metadata for each year
# filter by cpt code
code_regex = "(?:70450|70460|70470|70551|70552|70553)"
dfs = []
for i in range(1, key_length + 1):
    df = reader[f'cpt_partition_{i}']
    df = df[df['CPT'].astype(str).str.match(code_regex)]
    dfs.append(df)

# show dataframe, number of discharge summaries
df = pd.concat(dfs, axis=0, ignore_index=True)
print(len(df))
df.head()

# only include get 1 per patient, filtering for those that have ever gotten CT Head or MRI
cpt_pt_ids = set(df['BDSPPatientID'])
print(len(cpt_pt_ids))

# read in discharge summaries df
icd_discharge_df = pd.read_csv('2_discharge_summaries_df.csv')
icd_discharge_df.head()


print(type(icd_discharge_df['BDSPPatientID'][0]))
print(type(df['BDSPPatientID'][0]))

# make patient IDs same type
icd_discharge_df['BDSPPatientID'] = icd_discharge_df['BDSPPatientID'].astype(int).astype(str)

icd_discharge_pt_ids = set(icd_discharge_df['BDSPPatientID'])

# find the overlap between cpt matching patients and previously filtered patients
overlap = cpt_pt_ids.intersection(icd_discharge_pt_ids)
print(len(overlap))

1,113 patients who overlap.  Let's filter the icd_discharge filtered file to only include those patients

# show length (looks like it filtered out about 400 patients) and current df
prepped_df = icd_discharge_df[icd_discharge_df['BDSPPatientID'].isin(overlap)]
print(len(prepped_df))
prepped_df.head()

# save as csv
prepped_df.to_csv('3_full_filtered.csv', index=False, header=True, sep=',', na_rep='NA')
