==> newFlowCohortExtractionPipeline/MGB/icd_pos/4_cohort_selection.ipynb <==
# imports
import pandas as pd
from thunderpack import ThunderReader
from tqdm import tqdm
import re

# Goal
- Obtain random sample of patients
## Plan
- Randomize order of df
- Filter df to only include 1 row per patient
- 

# read in full filtered csv
full_filtered_df = pd.read_csv('3_full_filtered.csv')

# set of patient ids in full filtered df
discharged_patients = set(full_filtered_df['BDSPPatientID'])
print(len(discharged_patients))

# randomize order of df, 25 seed for reproducability
rand_filtered_df = full_filtered_df.sample(frac=1, random_state=25).reset_index(drop=True)
rand_filtered_df.head()

# keep only the first instance of patient id
rand_filtered_df = rand_filtered_df.drop_duplicates(subset='BDSPPatientID', keep='first')
print(len(rand_filtered_df))

# save as csv
rand_filtered_df.to_csv('4_rand_filtered_df.csv', index=False)
