==> newFlowCohortExtractionPipeline/MGB/icd_pos/5_note_text.ipynb <==
import pandas as pd
from thunderpack import ThunderReader
from tqdm import tqdm
import tarfile
import zipfile
import random
import os
import shutil
import re
tqdm.pandas()

# Goal of this notebook:
 - Add a column to the dataframe which includes the text of the note associated with the row
 - Plan:
    - Figure out how to open tar archive from here
    - Create a function that opens the tar archive, finds the file, and extracts the text
    - Make a pandas series from these text blocks and append to the initial dataframe

rand_filtered_df = pd.read_csv('4_rand_filtered_df.csv')
rand_filtered_df.head()

def convert_tar_to_zip(tar_path, zip_path):
    with tarfile.open(tar_path, 'r') as tar:
        with zipfile.ZipFile(zip_path, 'w') as zipf:
            for member in tqdm(tar.getmembers()):
                file_obj = tar.extractfile(member)
                if file_obj is not None:
                    zipf.writestr(member.name, file_obj.read())

# This will extract the year from the file name
def extract_year(file_name):
    match = re.search(r'_(\d{8})\.txt$', file_name)
    if match:
        return match.group(1)[:4]
    return None
def read_file_from_zip(base_path, file_name):
    year = extract_year(file_name)
    if year:
        zip_path = f'{base_path}/mgb_notes_{year}.zip'
        try:
            with zipfile.ZipFile(zip_path, 'r') as zipf:
                with zipf.open(f'{year}/{file_name}') as file:
                    content = file.read()
                    return content.decode('utf-8', errors = 'ignore')
        except (KeyError, FileNotFoundError):
            return None
    return None

year = 2013
for i in tqdm(range(0, 12)):
    tar_file_path = f'/home/jsearle/bigDrive/Dropbox/zz_EHR_Thunderpacks/MGB/MGB_Deidentified_Notes_March12th2024/mgb_notes_{year + i}.tar'
    zip_file_path = f'/home/jsearle/bigDrive/NAX/zipNotes/MGB/mgb_notes_{year + i}.zip'
    convert_tar_to_zip(tar_file_path, zip_file_path)

base_path = '/home/jsearle/bigDrive/NAX/zipNotes/MGB'

rand_filtered_df['text'] = rand_filtered_df['NoteTextFile'].progress_apply(lambda x: read_file_from_zip(base_path, x))

rand_filtered_df.to_csv('cohort_plus_text.csv', index=False, header=True, sep=',', na_rep='NA')

rand_filtered_df.head()
print(len(rand_filtered_df))

def check_word_count(text, threshold):
    word_count = len(text.split())
    return word_count > threshold

threshold = 500
rand_filtered_df['valid'] = rand_filtered_df['text'].apply(lambda x: check_word_count(x, threshold))
rand_filtered_df.head()

numValid = sum(rand_filtered_df['valid'])
print(numValid)

false_indeces = rand_filtered_df.index[~rand_filtered_df['valid']].tolist()
print(false_indeces)

final_icd_plus_df = rand_filtered_df.drop(index=false_indeces)
print(len(final_icd_plus_df))
final_icd_plus_df.head()

# write to csv
final_icd_plus_df.to_csv('5_final_icd_plus_df.csv', index=False)

final_icd_plus_df = pd.read_csv('5_final_icd_plus_df.csv')

# df cleanup
keepColumns = ['BDSPPatientID', 'DateNote',  'NoteTextFile', 'text']
clean_final_icd_plus_df = final_icd_plus_df[keepColumns]

clean_final_icd_plus_df.head()
